+++ pkg/modules/authorization_test.go
@@ -807,67 +807,3 @@ func TestNginxIngressController(t *testing.T) {
 		})
 	}
 }
-
-func TestAuthorizationCertManager(t *testing.T) {
-	tests := map[string]func(t *testing.T) (bool, bool, csmv1.ContainerStorageModule, ctrlClient.Client, utils.OperatorConfig){
-		"success - deleting": func(*testing.T) (bool, bool, csmv1.ContainerStorageModule, ctrlClient.Client, utils.OperatorConfig) {
-			customResource, err := getCustomResource("./testdata/cr_auth_proxy.yaml")
-			if err != nil {
-				panic(err)
-			}
-
-			tmpCR := customResource
-
-			cr := &appsv1.Deployment{
-				TypeMeta: metav1.TypeMeta{
-					Kind: "Deployment",
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "authorization-cert-manager",
-				},
-			}
-
-			sourceClient := ctrlClientFake.NewClientBuilder().WithObjects(cr).Build()
-			return true, true, tmpCR, sourceClient, operatorConfig
-		},
-		"success - creating": func(*testing.T) (bool, bool, csmv1.ContainerStorageModule, ctrlClient.Client, utils.OperatorConfig) {
-			customResource, err := getCustomResource("./testdata/cr_auth_proxy.yaml")
-			if err != nil {
-				panic(err)
-			}
-
-			tmpCR := customResource
-
-			sourceClient := ctrlClientFake.NewClientBuilder().WithObjects().Build()
-			return true, false, tmpCR, sourceClient, operatorConfig
-		},
-
-		"fail - wrong module name": func(*testing.T) (bool, bool, csmv1.ContainerStorageModule, ctrlClient.Client, utils.OperatorConfig) {
-			customResource, err := getCustomResource("./testdata/cr_powerscale_replica.yaml")
-			if err != nil {
-				panic(err)
-			}
-
-			tmpCR := customResource
-
-			sourceClient := ctrlClientFake.NewClientBuilder().WithObjects().Build()
-
-			return false, false, tmpCR, sourceClient, operatorConfig
-		},
-	}
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-
-			success, isDeleting, cr, sourceClient, op := tc(t)
-
-			err := AuthorizationCertManager(context.TODO(), isDeleting, op, cr, sourceClient)
-			if success {
-				assert.NoError(t, err)
-
-			} else {
-				assert.Error(t, err)
-			}
-
-		})
-	}
-}
